refactor(settings): remove abstraction over store
